pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "crypto-trading-control-plane"
version = "0.1.0"
description = "Local-first crypto trading control plane and paper trading lab"
requires-python = ">=3.11"
dependencies = [
  "cryptography>=44,<46",
  "Flask>=3.1,<4",
  "python-dotenv>=1.0,<2",
  "requests>=2.32,<3",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
